pyproject.toml
[tool.poetry]
name = "flasker"
version = "0.1.0"
description = ""
authors = ["inessochan <iness0chan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
flask-wtf = "^1.0.1"
flask-migrate = "^3.1.0"
flask-sqlalchemy = "^3.0.2"
flask-login = "^0.6.2"
email-validator = "^1.3.0"
flask-mailman = "^0.3.0"
pyjwt = "^2.6.0"
flask-moment = "^1.0.5"
flask-babel = "^2.0.0"
langdetect = "^1.0.9"
elasticsearch = "^8.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
